Intent_Classifier/backend/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
openai==1.93.1
sentence-transformers==5.0.0
transformers==4.53.1
nest-asyncio==1.6.0
pandas==2.3.1
numpy==2.3.1
requests==2.32.4
python-multipart==0.0.9
tqdm==4.67.1 